添加了Opengl上下文

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -71,7 +71,7 @@ namespace Hazel {
     {
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::Get();
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
         // 核心渲染函数，生成绘制数据
         ImGui::Render();
         ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());     //渲染
